backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-dateutil>=2.8.2
